[KT] Get rid of duplicates
Restructure headers
